krb5: Use static buffers for HMAC
Rather than mallocing storage in the HMAC routine, which is an
expensive operation when we're dealing with relatively small amounts
of space, use stack buffers for padding and iovecs. Merge the ipad and
opad buffers into a single space 'pad', so we only have one thing to
zero at the end.

NOTE: The buffer sizes here are arbitrarily chosen, and no attempt is
made to ensure that they are large enough!!! Not yet fit for public
consumption!